Created AutoValueEditor components (out of bits previously from CMDElementForm and SpecFormUpdateMixin) To be reused in CMD attribute form
