+++ src/scripts/app/mixins/SpecFormUpdateMixin.js
@@ -2,7 +2,6 @@
 
 var log = require('loglevel');
 var changeObj = require('../util/ImmutabilityUtil').changeObj;
-var update = require('../util/ImmutabilityUtil').update;
 var ComponentSpec = require('../service/ComponentSpec');
 /**
 * SpecFormUpdateMixin - Common functions and properties for the specification
@@ -25,26 +24,6 @@ var SpecFormUpdateMixin = {
   updateDocumentation: function(documentation) {
     log.trace("Update documentation", documentation);
     this.propagateValue("Documentation", documentation);
-  },
-
-  addAutoValueExpression: function() {
-    var spec = this.props.spec;
-    if($.isArray(spec.AutoValue)) {
-      this.propagateValue("AutoValue", update(spec.AutoValue, {$push: [""]}));
-    } else {
-      this.propagateValue("AutoValue", [""]);
-    }
-  },
-
-  updateAutoValueExpression: function(index, evt) {
-    log.debug("Update auto value expression", index, evt);
-    var newValue = evt.target.value;
-    this.propagateValue("AutoValue", update(this.props.spec.AutoValue, {$splice: [[index, 1, newValue]]}));
-  },
-
-  removeAutoValueExpression: function(index) {
-    log.debug("Remove auto value expression", index);
-    this.propagateValue("AutoValue", update(this.props.spec.AutoValue, {$splice: [[index, 1]]}));
   },
 
   addNewAttribute: function(onChange, evt) {
